Add configuration to determine whether to validate plugins to install
This commit adds a ValidatePluginsToInstall bool property to EphemeralClusterConfiguration
to determine whether the plugins to install should be validated for the version of Elasticsearch
before attempting to install. This can be useful to fail early in a scenario where a plugin will
be relied upon in a later operation but was not valid for installation. The default is true.