Use plugin bat when installing on Windows (#23)
This commit updates the InstallPlugins task to use the plugin batch file as the binary when installing plugins. Previously,
plugin installation used cmd as the binary, delegating to the batch file as an argument. This can cause issues where a plugin
writes warnings to stderr, such as ingest-attachmeent, which triggers the InstallPlugins task to throw an exception.

Remove IsMono property and replace with IsWindows property, using RuntimeInformation. Ultimately, this property is used to determine
the platform running on, specifically whether Windows or not.